移植bilibili视频合并(未完成), 修复模组更新可编程相关问题

+++ src/libs/updateGTNHMods.ts
@@ -2,8 +2,9 @@ import fs from "fs";
 import { argv } from "process";
 import { confirm, input } from "../utils/input.js";
 import { config } from "../data/index.js";
-import { Mods } from "../data/modsConfig/index.js";
+import { Mods } from "../data/Config/ModsConfig.js";
 import { Writable } from "stream";
+import chalk from "chalk";
 
 const modPath = config.modsConfig.modPath;
 
@@ -12,21 +13,36 @@ export const updateMods = async (mods: Mods[]) => {
     const oldFileName = `${mod.name}-${mod.version}.jar`;
     const response = await fetch(
       `https://api.github.com/repos/${mod.url}/releases/latest`,
+      {
+        headers: {
+          Authorization: `Bearer ${config.githubUserConfig.token}`,
+          "User-Agent": "jsutil by hirrao",
+        },
+      },
     );
     if (!response.ok) {
+      console.error(await response.json());
       console.log("获取失败");
       return -1;
     }
     const data = await response.json();
-    const newVersion = data.tag_name;
-    if (mod.version === newVersion) {
-      console.log(`${mod.name}已经是最新版本`);
+    const newVersionTag = data.tag_name as string;
+    const newVersionName = (() => {
+      if (mod.name === "programmablehatches") {
+        return newVersionTag.replace("v", "").split("-")[0];
+      } else return newVersionTag;
+    })();
+    if (mod.version === newVersionName) {
+      console.log(`${chalk.blue(mod.name)}已经是最新版本`);
       return 1;
     }
-    const newFileName = `${mod.name}-${newVersion}.jar`;
-    const download = await fetch(
-      `https://github.com/${mod.url}/releases/download/${newVersion}/${mod.name}-${newVersion}.jar`,
+    console.log(
+      `${chalk.blue(mod.name)}的最新版本是${newVersionTag}, 正在尝试更新`,
     );
+    const newFileName = `${mod.name}-${newVersionName}.jar`;
+    const fileUrl = `https://github.com/${mod.url}/releases/download/${newVersionTag}/${mod.name}-${newVersionName}.jar`;
+    const download = await fetch(fileUrl);
+    console.log(`正在下载  ${chalk.yellow(fileUrl)}`);
     const nodeStream = Writable.toWeb(fs.createWriteStream(newFileName));
     if (download.body) {
       await download.body.pipeTo(nodeStream);
@@ -38,11 +54,13 @@ export const updateMods = async (mods: Mods[]) => {
       return 0;
     }
     modPath.forEach((path) => {
-      fs.unlinkSync(`${path}/${oldFileName}`);
+      if (fs.existsSync(`${path}/${oldFileName}`)) {
+        fs.unlinkSync(`${path}/${oldFileName}`);
+      }
       fs.cpSync(newFileName, `${path}/${newFileName}`);
     });
+    mod.version = newVersionName;
     fs.unlinkSync(newFileName);
-    mod.version = newVersion;
   });
 };
 
@@ -57,7 +75,7 @@ const addMods = async () => {
   });
 };
 
-export const updateGTNHMods = async () => {
+export const checkUpdateMods = async () => {
   if (argv[3]) {
     addMods();
   } else {
@@ -65,3 +83,5 @@ export const updateGTNHMods = async () => {
   }
   return 0;
 };
+
+export default checkUpdateMods;
